Estimate the soil sorption coefficient in two different ways

Estimation via direct or indirect logarithmic linear regression including pH, organic carbon and soil concentration. (log10 is used for all variables)

Kd_regression(
  constant,
  beta_ph,
  beta_org,
  beta_conc,
  regType = "direct",
  pH,
  org_c,
  conc
)

Arguments

constant

Log constant in linear Kd regression

beta_ph, beta_org, beta_conc

Log coefficient for pH, organic carbon and soil concentration in log-linear Kd regression

regType

Chacter specifying the way the K_d value is calculated. Either "direct" which means that the regression directly leads to the sorption coefficient or"indirect" which means the regression estimates a concentration in pore water and the sorption coefficient is subsequently be calculated by the quotient of soil and pore water concentration.

pH

pH value

org_c

Organic carbon content in percent.

conc

Soil concentration in mg/kg

Value

A vector of sorption coefficients in L/kg

Details

The constant value, the regression parameter (beta) and the soil characteristics can be single numeric values or vectors of the same length.