PECs for soil, porewater and human consumption over a long period of time

Based on the input data provided in Excel sheets, 3 different PECs are calculated. The number of years and number of agricultural fields with different properties (Number of Monte Carlo cycles) can be chosen.

longterm_PEC(
  dat,
  info,
  years,
  nFields,
  use_mixing_factor = FALSE,
  food_only = TRUE,
  growing_period = 180,
  t_res = 1,
  traceBackVariables = FALSE
)

Arguments

dat

List with all the input variables. This list is produced by function read_fcr_input() from the Excel sheets.

info

he table containing additional substance information loaded with additional_substanc_info()

years

Years of fertilizer application

nFields

Number of Monte Carlo cycles

use_mixing_factor

Not working yet! If TRUE, porewater is diluted by pollutant free groundwater for a more realistic estimation of risks in groundwater. However the TGD approach assumes porewater = groundwater for the assessment.

food_only

If TRUE, the predicted human consumption via food is multiplied by 2 to compensate the disregard of water consumption

growing_period

Numeric value specifiyng the number of days with plant growth

t_res

Temporal resolution to be returned in the output matrix (no effect on the calculation). The default is 1 (every year) which can lead to a very large output matrix.

traceBackVariables

If TRUE, the variables of every simulation year are returned. Note: this can lead to very large lists.

Value

List with

table for all PEC types per year,

table for concentration course in top soil in the predefined temporal resolution,

optional: If traceBackVariables TRUE: list of distributed input variables for each year, if traceBackVariables FALSE: Input variables of the first year including initial situations for human consumption and inital concentration in porewater