Distribution derived by comparable circumstances

This function combines a uniform distribution from minimum to maximum with the upper and lower tail of a normal distribution.

rderived(n, min, max, a = -Inf)

Arguments

n

Number of samples to be drawn

min

Minimum value of normal comparables

max

Maximum value of normal comparables

a

Minimum of distribution (Default is -Inf)

Value

numeric vector of length n with randomly drawn values according to the distribution

Details

95%, 2.5% and 2.5% of all drawn values originate from the uniform distribution, the lower tail, and the upper tail of a normal distribution, respectively. The normal distribution is characterized by the mean of minimum and maximum. The standard deviation is derived from the 2.5% and 97.5% quantile definied by the minimum and maximum values. (sd = (max - mean) / 1.959963)