PECs for soil, porewater and human consumption over a long period of time

Based on the input data provided in Excel sheets, 3 different PECs are calculated. The number of years and number of agricultural fields with different properties (Number of Monte Carlo cycles) can be chosen.

Usage

longterm_PEC(
  dat,
  info,
  years,
  nFields,
  use_mixing_factor = FALSE,
  PNECwater_c_i = FALSE,
  food_only = TRUE,
  growing_period = 180,
  t_res = 1,
  traceBackVariables = FALSE,
  keep_c_course = TRUE
)

Arguments

dat

A List with all the input variables. This list is produced by function read_fcr_input() from the Excel sheets.

info

A table containing additional substance information loaded with read_fcr_input()

years

Years of fertilizer application

nFields

Number of Monte Carlo cycles

use_mixing_factor

Not working yet! If TRUE, porewater is diluted by pollutant free groundwater for a more realistic estimation of risks in groundwater. However the TGD approach assumes porewater = groundwater for the assessment.

PNECwater_c_i

If TRUE, the initial concentration in soil is derived by a RQ = 1 for groundwater risk assessment (PEC_porewater = PNEC_porewater). Thus, the inital concentration is defined by the applied PNEC for groundwater.

food_only

If TRUE, the predicted human consumption via food is multiplied by 2 to compensate the disregard of water consumption

growing_period

Numeric value specifiyng the number of days with plant growth

t_res

Temporal resolution in days to be returned in the output matrix (no effect on the calculation). The default is 1 (every day) which can lead to a very large output matrix. The maximal reslution is 365 days, since every year is calculated seperately.

traceBackVariables

If TRUE, the variables of every simulation year are returned. Note: this can lead to very large lists.

keep_c_course

If FALSE, the course of concentration is not part of the output. In that case variabel t_res is meaningless.

Value

List with

table for all PEC types per year,

table for concentration course in top soil in the predefined temporal resolution,

optional: If traceBackVariables TRUE: list of distributed input variables for each year, if traceBackVariables FALSE: Input variables of the first year including initial situations for human consumption and inital concentration in porewater